fixed money

+++ src/main/java/com/techelevator/Candy.java
@@ -1,9 +1,11 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Candy extends Product {
     private String message = "Munch Munch, Yum!";
 
-    public Candy(String name, double cost, String slot) {
+    public Candy(String name, BigDecimal cost, String slot) {
         super(name, cost, slot);
     }
