+++ src/main/java/com/techelevator/Chip.java
@@ -1,10 +1,12 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Chip extends Product {
 
     private String message = "Crunch Crunch, Yum!";
 
-    public Chip(String name, double cost, String slot) {
+    public Chip(String name, BigDecimal cost, String slot) {
         super(name, cost, slot);
     }
