+++ src/main/java/com/techelevator/Drink.java
@@ -1,10 +1,12 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Drink extends Product {
 
     private String message = "Glug glug, Yum!";
 
-    public Drink(String name, double cost, String slot) {
+    public Drink(String name, BigDecimal cost, String slot) {
         super(name, cost, slot);
     }
