+++ src/main/java/com/techelevator/Gum.java
@@ -1,10 +1,12 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Gum extends Product {
     private String message = "Chew Chew, Yum!";
 
 
-    public Gum (String name, double cost, String slot) {
+    public Gum (String name, BigDecimal cost, String slot) {
         super(name, cost, slot);
     }
